icu_ai/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
openai==1.59.5
anthropic==0.40.0

python-dotenv==1.0.0
SQLAlchemy==2.0.36
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
requests==2.31.0
gunicorn==21.2.0
psycopg[binary]==3.3.2

# Authentication
PyJWT==2.10.1
bcrypt==5.0.0
python-dateutil==2.9.0
